+++ security/pkg/server/ca/authenticate/kube_jwt.go
@@ -15,8 +15,6 @@
 package authenticate
 
 import (
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
 	"strings"
 
@@ -24,8 +22,11 @@ import (
 	"google.golang.org/grpc/metadata"
 	"k8s.io/client-go/kubernetes"
 
+	"istio.io/istio/security/pkg/util"
+
 	"istio.io/istio/pkg/security"
 	"istio.io/istio/security/pkg/k8s/tokenreview"
+	"istio.io/istio/security/pkg/util"
 )
 
 const (
@@ -54,40 +55,6 @@ type KubeJWTAuthenticator struct {
 }
 
 var _ Authenticator = &KubeJWTAuthenticator{}
-
-type jwtPayload struct {
-	// Aud is JWT token audience - used to identify 3p tokens.
-	// It is empty for the default K8S tokens.
-	Aud []string `json:"aud"`
-}
-
-// isK8SUnbound detects if the token is a K8S unbound token.
-// It is a regular JWT with no audience and expiration, which can
-// be exchanged with bound tokens with audience.
-//
-// This is used to determine if we check audience in the token.
-// Clients should not use unbound tokens except in cases where
-// bound tokens are not possible.
-func isK8SUnbound(jwt string) bool {
-	jwtSplit := strings.Split(jwt, ".")
-	if len(jwtSplit) != 3 {
-		return false // unbound tokens are valid JWT
-	}
-	payload := jwtSplit[1]
-
-	payloadBytes, err := base64.RawStdEncoding.DecodeString(payload)
-	if err != nil {
-		return false // unbound tokens are valid JWT
-	}
-
-	structuredPayload := &jwtPayload{}
-	err = json.Unmarshal(payloadBytes, &structuredPayload)
-	if err != nil {
-		return false
-	}
-
-	return len(structuredPayload.Aud) == 0
-}
 
 // NewKubeJWTAuthenticator creates a new kubeJWTAuthenticator.
 func NewKubeJWTAuthenticator(client kubernetes.Interface, clusterID string,
@@ -128,7 +95,7 @@ func (a *KubeJWTAuthenticator) Authenticate(ctx context.Context) (*Caller, error
 	// If 'Require3PToken' is set - we will also set the audiences field, forcing the check.
 	// If Require3P is not set - and token does not have audience - we will
 	// tolerate the unbound tokens.
-	if !isK8SUnbound(targetJWT) || security.Require3PToken.Get() {
+	if !util.IsK8SUnbound(targetJWT) || security.Require3PToken.Get() {
 		aud = security.TokenAudiences
 		// TODO: check the audience from token, no need to call
 		// apiserver if audience is not matching. This may also
